rebuild_plan/docker/requirements.txt
# GTK and Cairo bindings
# These are handled by the system packages and installed separately in Dockerfile
# PyGObject>=3.42.0
# pycairo>=1.20.0

# Image processing
Pillow>=9.0.0

# Gemini AI integration
google-generativeai>=0.3.0

# Configuration management
toml>=0.10.2
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
# pytest-gtk>=0.2.0 - Package not available
flake8>=6.0.0
mypy>=1.0.0
black>=23.0.0
isort>=5.12.0

# Debugging tools
debugpy>=1.6.0
psutil>=5.9.0

# Utilities for testing
faker>=18.0.0
hypothesis>=6.80.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Networking and API
requests>=2.28.0
httpx>=0.24.0